Moves mapper functions to separate files
